listUser() 서버통신 로직을 HomeViewModel로 분리
